store inelastic scattering quantities as MultiDArray. Changed inelastic sampling to rejection sample outgoing bin, then weight for variations within the bin.

+++ src/spectrum/GR1DSpectrumArray.h
@@ -99,7 +99,7 @@ public:
 
 		int nu_bin = data.axes[nuGridIndex].bin(nu);
 		nu_bin = max(nu_bin, 0);
-		nu_bin = min(nu_bin, data.axes[nuGridIndex].size()-1);
+		nu_bin = min(nu_bin, (int)data.axes[nuGridIndex].size()-1);
 		indices[nuGridIndex] = nu_bin;
 
 		indices[momGridIndex] = 0;
